add dynamic route and finished all functionality (hopefully) ;) (yay)
