+++ src/app/api/file/[fileId]/route.schema.ts
@@ -26,20 +26,24 @@ export const File = z.object({
 
 export const FileResponse = z.discriminatedUnion("code", [
   z.object({
-    code: z.literal("SUCCESS"),
-    message: z.literal("File successfully added"),
-  }),
-  z.object({
-    code: z.literal("INVALID_FILE"),
-    message: z.literal("Invalid file added"),
-  }),
-  z.object({
     code: z.literal("UNKNOWN"),
     message: z.literal("Unknown error received"),
   }),
   z.object({
     code: z.literal("NOT_AUTHORIZED"),
     message: z.literal("Senior not assigned to user"),
+  }),
+  z.object({
+    code: z.literal("SUCCESS_UPDATE"),
+    message: z.literal("File successfully updated"),
+  }),
+  z.object({
+    code: z.literal("SUCCESS_DELETE"),
+    message: z.literal("File successfully deleted"),
+  }),
+  z.object({
+    code: z.literal("INVALID_URL"),
+    message: z.literal("Invalid file ID parsed from url"),
   }),
   z.object({
     code: z.literal("NO_SENIOR"),
