+++ src/app/api/file/route.client.ts
@@ -26,31 +26,3 @@ export const createFile = async (request: IRequest) => {
 
   return FileResponse.parse(json);
 };
-
-export const updateFile = async (request: IRequest, id?: string) => {
-  const { body, ...options } = request;
-
-  const response = await fetch("/api/file/", {
-    method: "PATCH",
-    body: JSON.stringify(body),
-    ...options,
-  });
-
-  const json = await response.json();
-
-  return FileResponse.parse(json);
-};
-
-export const deleteFile = async (request: IRequest) => {
-  const { body, ...options } = request;
-
-  const response = await fetch("/api/file", {
-    method: "DELETE",
-    body: JSON.stringify(body),
-    ...options,
-  });
-
-  const json = await response.json();
-
-  return FileResponse.parse(json);
-};
